// frontend/src/api/model/githubCom4H1RZooraInternalDomainGatewayName.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainGatewayName = typeof GithubCom4H1RZooraInternalDomainGatewayName[keyof typeof GithubCom4H1RZooraInternalDomainGatewayName];


export const GithubCom4H1RZooraInternalDomainGatewayName = {
  GatewayZarinpal: 'zarinpal',
  GatewayManual: 'manual',
} as const;